flask/requirements.txt
Flask==2.2.5
pymysql
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
mysql-connector-python